@@ -12,20 +12,12 @@ func btoi(b bool) int {
 	return 0
 }
 
-// max returns the greater of two integers.
-func max(a, b int) int { //nolint:predeclared
-	if a > b {
-		return a
+// bton converts a boolean to a specific integer, n if true, 0 if false.
+func bton(b bool, n int) int {
+	if b {
+		return n
 	}
-	return b
-}
-
-// min returns the smaller of two integers.
-func min(a, b int) int { //nolint:predeclared
-	if a < b {
-		return a
-	}
-	return b
+	return 0
 }
 
 // sum returns the sum of all integers in a slice.
