@@ -33,7 +33,6 @@ func ExampleLeaf_SetHidden() {
 	// │   └── Quux
 	// │       └── Hello!
 	// └── Baz
-	//
 }
 
 func ExampleNewLeaf() {
@@ -64,7 +63,6 @@ func ExampleNewLeaf() {
 	// │   │       └── leaves
 	// │   └── Quuux
 	// └── Baz
-	//
 }
 
 func ExampleLeaf_SetValue() {
@@ -86,14 +84,14 @@ func ExampleLeaf_SetValue() {
 	glossier.SetValue("Il Makiage")
 	fmt.Println(ansi.Strip(t.String()))
 	// Output:
-	//⁜ Makeup
-	//├── Il Makiage
-	//├── Fenty Beauty
-	//│   ├── Gloss Bomb Universal Lip Luminizer
-	//│   ╰── Hot Cheeks Velour Blushlighter
-	//├── Nyx
-	//├── Mac
-	//╰── Milk
+	// ⁜ Makeup
+	// ├── Il Makiage
+	// ├── Fenty Beauty
+	// │   ├── Gloss Bomb Universal Lip Luminizer
+	// │   ╰── Hot Cheeks Velour Blushlighter
+	// ├── Nyx
+	// ├── Mac
+	// ╰── Milk
 }
 
 // Tree Examples
@@ -144,6 +142,5 @@ func ExampleTree_SetHidden() {
 	// │   ├── Qux
 	// │   └── Quuux
 	// └── Baz
-	//
 	fmt.Println(tr.String())
 }
