@@ -12,9 +12,7 @@ type whitespace struct {
 	style Style
 }
 
-// newWhitespace creates a new whitespace renderer. The order of the options
-// matters, if you're using WithWhitespaceRenderer, make sure it comes first as
-// other options might depend on it.
+// newWhitespace creates a new whitespace renderer.
 func newWhitespace(opts ...WhitespaceOption) *whitespace {
 	w := &whitespace{}
 	for _, opt := range opts {
